# public/api-docs/swagger.yaml
swagger: "2.0"
host: festivals-app.herokuapp.com
basePath: /api
info:
  version: 1.0.0
  title: festivals
...
schemes:
  - https
produces:
  - application/vnd.festivals.v1+json
  - application/vnd.festivals.v1+xml
security:
  - oauth2token:
    - public
paths:
...
  "/festivals/{id}/events":
    get:
      tags:
        - events
      summary: Get festival events
      description: Get festival events collection
      operationId: getFestivalEvents
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - events:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: name
          description: name
          required: false
          type: string
          in: query
        - name: place
          description: place
          required: false
          type: string
          in: query
        - name: startAt
          description: startAt
          required: false
          type: string
          in: query
        - name: finishAt
          description: finishAt
          required: false
          type: string
          in: query
        - name: updatedAt
          description: updatedAt
          required: false
          type: string
          in: query
        - name: category
          description: category
          required: false
          type: string
          in: query
        - name: limit
          description: limit
          required: false
          type: integer
          in: query
          default: 100
        - name: offset
          description: offset
          required: false
          type: integer
          in: query
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/EventsCollectionResponse"
        "404":
          description: Not found
    post:
      tags:
        - events
      summary: Create festival events
      description: Create festival events
      operationId: createFestivalEvent
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - events:create
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - in: body
          name: festivalEventRequest
          description: Festival event object
          required: true
          schema:
            $ref: "#/definitions/EventRequest"
      responses:
        "201":
          description: successful operation
          schema:
            $ref: "#/definitions/EventResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Festival not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
  "/festivals/{id}/events/{event.id}":
    get:
      tags:
        - events
      summary: Get festival event
      description: Get festival event collection
      operationId: getFestivalEvent
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - events:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: event.id
          in: path
          description: Event id
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/EventResponse"
    delete:
      tags:
        - events
      summary: Delete festival event
      description: Delete festival events
      operationId: deleteFestivalEvent
      security:
        - oauth2token:
          - events:delete
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: event.id
          in: path
          description: Event id
          required: true
          type: string
      responses:
        "204":
          description: successful operation
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      tags:
        - events
      summary: Update festival events
      description: Update festival events
      operationId: updateFestivalEvent
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - events:update
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: event.id
          in: path
          description: Event id
          required: true
          type: string
        - in: body
          name: festivalEventRequest
          description: Festival event object
          required: true
          schema:
            $ref: "#/definitions/EventRequest"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/EventResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
...
  "/festivals/{id}/categories/{category.id}":
    get:
      tags:
        - categories
      summary: Get festival category
      description: Get festival category
      operationId: getFestivalCategory
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - categories:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: category.id
          in: path
          description: Category id
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/CategoryResponse"
        "404":
          description: Not found
    delete:
      tags:
        - categories
      summary: Delete festival category
      description: Delete festival category
      operationId: deleteFestivalCategory
      security:
        - oauth2token:
          - categories:delete
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: category.id
          in: path
          description: Category id
          required: true
          type: string
      responses:
        "204":
          description: successful operation
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      tags:
        - categories
      summary: Update festival category
      description: Update festival category
      operationId: updateFestivalCategory
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - categories:update
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: category.id
          in: path
          description: Category id
          required: true
          type: string
        - in: body
          name: festivalCategoryRequest
          description: Festival category object
          required: true
          schema:
            $ref: "#/definitions/CategoryRequest"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/CategoryResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
...
  "/festivals/{id}/places/{place.id}":
    get:
      tags:
        - places
      summary: Get festival place
      description: Get festival place
      operationId: getFestivalPlace
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - places:get
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: place.id
          in: path
          description: Place id
          required: true
          type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/PlaceResponse"
        "404":
          description: Festival not found
    delete:
      tags:
        - places
      summary: Delete festival place
      description: Delete festival place
      operationId: deleteFestivalPlace
      security:
        - oauth2token:
          - places:delete
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: place.id
          in: path
          description: Place id
          required: true
          type: string
      responses:
        "204":
          description: successful operation
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    put:
      tags:
        - places
      summary: Update festival place
      description: Update festival place
      operationId: updateFestivalPlace
      consumes:
        - application/vnd.festivals.v1+json
      produces:
        - application/vnd.festivals.v1+json
      security:
        - oauth2token:
          - places:update
      parameters:
        - name: id
          in: path
          description: Festival id
          required: true
          type: string
        - name: place.id
          in: path
          description: Place id
          required: true
          type: string
        - in: body
          name: festivalPlaceRequest
          description: Festival place object
          required: true
          schema:
            $ref: "#/definitions/PlaceRequest"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/PlaceResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "406":
          description: The request you made is not acceptable
        "422":
          description: Invalid parameter value
securityDefinitions:
  oauth2token:
    type: oauth2
    flow: implicit
    authorizationUrl: https://vehicle-history.io/oauth/authorization
    scopes:
      user: user auth scope
definitions:
  CoordinatesRequest:
    properties:
      lat:
        type: number
        format: double
        description: lat
      lon:
        type: number
        format: double
        description: lon
    description: Geo coordinates
    required:
      - lat
      - lon
  CoordinatesResponse:
    properties:
      lat:
        type: number
        format: double
        description: lat
      lon:
        type: number
        format: double
        description: lon
    description: Geo coordinates
    required:
      - lat
      - lon
  AuthorResponse:
    properties:
      name:
        type: string
        description: name
      organization:
        type: string
        description: organization
    description: Author
    required:
      - name
      - organization
...
  DurationResponse:
    properties:
      startAt:
        type: string
        description: start at
      finishAt:
        type: string
        description: finish at
      periodMs:
        type: integer
        description: periodMs
    description: Duration response
    required:
      - startAt
      - finishAt
      - periodMs
  DurationRequest:
    properties:
      startAt:
        type: string
        description: start at
      finishAt:
        type: string
        description: finish at
    description: Duration request
    required:
      - startAt
      - finishAt
...
  MainImageResponse:
    properties:
      small:
        type: string
        description: Image small url
      medium:
        type: string
        description: Image medium url
      large:
        type: string
        description: Image large url
    description: Festival response
  ImageRequest:
    properties:
      url:
        type: string
        description: Image url
      content:
        type: string
        description: Image content
      order:
        type: integer
        description: Image order
    description: Festival response
    required:
      - order
  PlaceResponse:
    properties:
      id:
        type: string
        description: Festival place id
      name:
        type: string
        description: Festival place name
      parent:
        type: string
        description: Festival place parent
      openingTimes:
        type: array
        description: Festival place opening times
        items:
          $ref: "#/definitions/DurationResponse"
      coordinates:
        $ref: "#/definitions/CoordinatesResponse"
        description: Coordinates
      createdAt:
        type: string
        description: Festival place createdAt
      updatedAt:
        type: string
        description: Festival place updatedAt
    description: Festival response
    required:
      - id
      - name
      - parent
      - openingTimes
      - createdAt
      - updatedAt
  PlaceRequest:
    properties:
      name:
        type: string
        description: Festival place name
      parent:
        type: string
        description: Festival place parent
      openingTimes:
        type: array
        description: Festival place opening times
        items:
          $ref: "#/definitions/Duration"
      coordinates:
        $ref: "#/definitions/CoordinatesRequest"
        description: Coordinates
    description: Festival response
    required:
      - name
      - parent
      - openingTimes
  EventPlaceResponse:
    properties:
      id:
        type: string
        description: Place id
      name:
        type: string
        description: Place name
      breadcrumbs:
        type: array
        description: Place breadcrumbs
        items:
          $ref: "#/definitions/PlaceResponse"
      openingTimes:
        type: array
        description: Place opening times
        items:
          $ref: "#/definitions/DurationResponse"
      coordinates:
        $ref: "#/definitions/CoordinatesResponse"
        description: Coordinates
    description: Festival response
    required:
      - id
      - name
      - breadcrumbs
      - openingTimes
  CategoryRequest:
    properties:
      parent:
        type: string
        description: Festival category parent id
      name:
        type: string
        description: Festival category name
    description: Festival category request
    required:
      - parent
      - name
  CategoryResponse:
    properties:
      id:
        type: string
        description: Category id
      name:
        type: string
        description: Category name
      parent:
        type: string
        description: Category parent
      createdAt:
        type: string
        description: Category createdAt
      updatedAt:
        type: string
        description: Category updatedAt
    description: Festival category response
    required:
      - id
      - name
      - parent
      - createdAt
      - updatedAt
  EventCategoryResponse:
    properties:
      id:
        type: string
        description: Event category id
      name:
        type: string
        description: Event category name
      breadcrumbs:
        type: array
        description: Event category breadcrumbs
        items:
          $ref: "#/definitions/CategoryResponse"
    description: Festival category response
    required:
      - id
      - name
      - breadcrumbs
  EventRequest:
    properties:
      name:
        type: string
        description: Event name
      description:
        type: string
        description: Event description
      tags:
        type: array
        items:
          type: string
        description: Event tags
      metadata:
        type: array
        items:
          type: string
        description: Event metadata
      images:
        type: array
        items:
          $ref: "#/definitions/ImageRequest"
        description: Event images
      duration:
        $ref: "#/definitions/DurationRequest"
        description: Event duration
      place:
        type: string
        description: Event place
      category:
        type: string
        description: Event category
      authors:
        type: array
        items:
          $ref: "#/definitions/AuthorResponse"
        description: Event authors
    description: Event request
    required:
      - name
      - description
      - duration
      - place
      - category
  EventResponse:
    properties:
      id:
        type: string
        description: Event id
      name:
        type: string
        description: Event name
      description:
        type: string
        description: Event description
      tags:
        type: array
        items:
          type: string
        description: Event tags
      mainImage:
        $ref: "#/definitions/MainImageResponse"
        description: Event main image
      duration:
        $ref: "#/definitions/DurationResponse"
        description: Event duration
      place:
        $ref: "#/definitions/EventPlaceResponse"
        description: Event place
      category:
        $ref: "#/definitions/EventCategoryResponse"
        description: Event category
      authors:
        type: array
        items:
          $ref: "#/definitions/AuthorResponse"
        description: Event authors
      createdAt:
        type: string
        description: Event createdAt
      updatedAt:
        type: string
        description: Event updatedAt
    description: Event response
    required:
      - id
      - name
      - description
      - duration
      - place
      - category
      - createdAt
      - updatedAt
...
  EventsCollectionResponse:
    properties:
      total:
        type: integer
        description: Total festival events
      events:
        type: array
        items:
          $ref: "#/definitions/EventResponse"
    description: Events collection
    required:
      - total
      - events
...